// src/pages/src/lib/database.types.ts
export interface User {
  user_id: string;
  email: string | null;
  password_hash: string | null;
  first_name: string;
  last_name: string;
  phone1: string | null;
  phone2: string | null;
  school: string | null;
  grade: string | null;
  dob: string | null;
  user_name: string | null;
  address: string | null;
  emergency_contact: string | null;
  medical_condition: string | null;
  status: 'Active' | 'Inactive' | 'Graduated' | 'Left';
  created_at: string;
}